% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list.R
\docType{data}
\name{list}
\alias{list}
\title{a structure variable with class attribute of result}
\format{\preformatted{Class 'result'  logi NA
}}
\usage{
list
}
\description{
\code{list} is a structure variable with class attribute of \code{result}.
}
\keyword{datasets}

