% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list.R
\name{[<-.result}
\alias{[<-.result}
\title{Extract variables from named elements of a list object}
\usage{
S3 method for class 'result'
list[] <- value
list[x, ] <- value
list[, y] <- value
list[x, y, ..., all = FALSE] <- value
}
\arguments{
\item{x,y}{variable names. List elements at corresponding index are extracted and then
assigned to those variables.}

\item{...}{empty strings for positions of list elements without interest. By default
these elements are not extracted to corresponding variables with same names.
Setting \code{all} to TRUE can also extract those elements without interest.}

\item{value}{a list object with named elements.}

\item{list}{an exported structure variable with class attribute of \code{result}.}

\item{all}{whether extract all named elements to variables. The default is FALSE
# except for the case without specifying any arguments, in which all elements
are extracted to corresponding variables.}
}
\value{
named variables such as \code{x}, \code{y} and variables with element names.
}
\description{
Operator acting on a list with named elements to extract list elements of
interest to named variables.
}
\examples{
# unpack results from function returning multiple values
# define a function returning two values
fun <- function() list(a = 1, b = 2)

# implicitly extract all variabes without renaming
list[] <- fun()  # return a = 1, b = 2

# explicitly extract all variables without renaming
list[, , all = TRUE] <- fun()  # return a = 1, b = 2

# no variables extracted
list[, , all = FALSE] <- fun()

# only extract the first variable and assign it to x. note that by default only
# element of interest is extracted and renamed.
list[x, ] <- fun()  # return x = 1

# set all to TRUE to extract all elements.
list[x, , all = TRUE] <- fun()  # return x = 1, b = 2

# only extract the second variable and assign it to y
list[, y] <- fun()  # return y = 2

# explicitly extract all variables with renaming
list[x, y] <- fun()  # return x = 1, y = 2

# parameter all ignored
list[x, y, all = TRUE] <- fun()  # return x = 1, y = 2
}
\author{
Jun Cai (\email{cai-j12@mails.tsinghua.edu.cn}), Ph.D. student from
Center for Earth System Science, Tsinghua University
}

